Pushed initial class
Pushed from local repository the initial script for this project. Created a landslide class, as well as created a method to visualize the landslides that occurred during the times recorded on a map, with respective visualizations to size. Visualizations will soon be added to the 'pictures' folder. Created constructors and formatting methods as well, which allows for the project to receive input and formats the titles and various axes.